condensed import lines / seperate component from pages
