+++ client/src/components/Copyright/index.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Typography, Link } from '@material-ui/core';
+
+function Copyright() {
+    return (
+      <Typography variant="body2" color="textSecondary" align="center">
+        {'Copyright © '}
+        <Link color="inherit" href="https://material-ui.com/">
+          Your Website
+        </Link>{' '}
+        {new Date().getFullYear()}
+        {'.'}
+      </Typography>
+    );
+}
+
+export default Copyright;
