+++ client/src/pages/Home.js
@@ -1,10 +1,6 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import GridList from '@material-ui/core/GridList';
-import GridListTile from '@material-ui/core/GridListTile';
-import GridListTileBar from '@material-ui/core/GridListTileBar';
-import ListSubheader from '@material-ui/core/ListSubheader';
-import IconButton from '@material-ui/core/IconButton';
+import { GridList, GridListTile, GridListTileBar, ListSubheader, IconButton } from '@material-ui/core'
 import InfoIcon from '@material-ui/icons/Info';
 import tileData from './tileData';
